tasks/web/snow_hell/deploy/service/requirements.txt
fastapi==0.62.0
uvicorn==0.11.5
python-multipart==0.0.5
aiofiles
jinja2
python_jose==3.1.0
